universal7904: Don't include mount skipping configuration
and so don't skip mounting /product, which the config causes, and make a
separate product partition usable.

The mount skipping config is intended to be included in GSIs, to not
use device-specific product from a product partition due to possible
issues.

We're not building a GSI while using this tree, and installable GSIs
should remain usable without including the config in system images as
part of device-specific builds. So why would this be needed?

Change-Id: I3b375e6a4cade5796cc9b0b44096e10df6a90d8f