universal7904: Remove camera wrapper
It doesn't seem to solve the problem it was originally intended to
solve.

It's also crashing with the libexynoscamera3 blob on A40, due to a small
bug in that blob which means it returns null after a meaningless error:
10-15 09:43:11.898  6103  6103 E SecCameraUtil: ERR(getSensorName[1124]): Unknown sensor(58), Return NULL
and there's a mess up the wrapper with a null pointer dereference, while
the blob can just continue on its own with the result that the cameras
all work fine.

Change-Id: Ia2346716f8482c6c40fa7794085344de0b4386b5